#305 splitt Auth trait file in two
