+++ src/main/scala/no/ndla/listingapi/auth/Role.scala
@@ -5,12 +5,12 @@
  * See LICENSE
  */
 
-package no.ndla.listingapi.service
+package no.ndla.listingapi.auth
 
 import no.ndla.listingapi.model.api.AccessDeniedException
 import no.ndla.network.AuthUser
 
-trait AuthenticationRole {
+trait Role {
 
   val authRole: AuthRole
 
@@ -19,25 +19,6 @@ trait AuthenticationRole {
       if (!AuthUser.hasRole(role))
         throw new AccessDeniedException("User is missing required role to perform this operation")
     }
-  }
-
-}
-
-trait AuthenticationUser {
-
-  val authUser: AuthUser
-
-  class AuthUser {
-
-    def id(): String = {
-      if (AuthUser.get.isEmpty || AuthUser.get.get.isEmpty) {
-        throw new AccessDeniedException(("User id required to perform this operation"))
-      } else {
-        return AuthUser.get.get
-      }
-
-    }
-
   }
 
 }
